refactoring
Split data processing and keras utils
